% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_shannon.R
\name{my_shannon}
\alias{my_shannon}
\title{My shannon diversity function}
\usage{
my_shannon(x)
}
\arguments{
\item{x}{A vector of numeric values representing species abundance in
a communty}
}
\value{
Shannon diversiy value: sum -pi*log(pi)
}
\description{
This is the best function ever! Thanks, Shannon.
}
\examples{
com <- c(1,4,7,8)
my_shannon(com) == 1.21
}
